// apps/formule1/src/app/constants/app.enums.ts
// all enums used in 1 place
export enum CountdownType {
  Big = 'big',
  Large = 'large',
  Medium = 'medium',
  Small = 'small'
}

export enum SessionNames {
  FirstPractice = 'FirstPractice',
  SecondPractice = 'SecondPractice',
  ThirdPractice = 'ThirdPractice',
  Qualifying = 'Qualifying',
  Sprint = 'Sprint',
  Race = 'Race',
  None = ''
}